+++ src/Domain.LinnApps/StockMove/Models/AvailableStock.cs
@@ -6,7 +6,7 @@
     {
         public string PartNumber { get; set; }
 
-        public int QuantityAvailable { get; set; }
+        public decimal QuantityAvailable { get; set; }
 
         public DateTime StockRotationDate { get; set; }
