+++ src/Resources/AvailableStockResource.cs
@@ -4,7 +4,7 @@
     {
         public string PartNumber { get; set; }
 
-        public int QuantityAvailable { get; set; }
+        public decimal QuantityAvailable { get; set; }
 
         public string StockRotationDate { get; set; }
